Animated Sprites and Sprite Scripts
The scripts are fairly basic. There are other things that are being added in a couple of different scenes. There is one issue, enemyFred, that I can't quite figure out. That code is what breaks no matter what I do. I get it to work and then it breaks. It's just the turn that doesn't seem to want to work.